# Cauchy distribution: optimise sampling