@MaximoB this presumably uses 1 bit less precision but is quite a bit faster
```
# previously
test distr_cauchy                ... bench:      48,639 ns/iter (+/- 1,411) = 164 MB/s                                                 
# now
test distr_cauchy                ... bench:      36,224 ns/iter (+/- 654) = 220 MB/s
```
Optimisers are weird, but as I guessed, it seems we get a free subtraction operation when using `Open01`.

I considered checking that the result is finite with a loop, but I'm fairly sure it must be anyway since π is irrational. Do you agree?