@vks 2018-05-30
> I considered checking that the result is finite with a loop, but I'm fairly sure it must be anyway since π is irrational.

I think you are right that `tan` produces finite results for finite input. The only thing I can think of that could go wrong is that `r * pi / 2.` rounds such that you end up on the other sight of the singularity, changing the sign.