@dhardy 2018-05-30
> The performance difference is because of using Open01, and removing the loop.

You're right; turns out removing the loop makes most of the difference. Moving everything inside the loop also performs well on my CPU, but I think we can just drop the loop since [tan(π/2) rounds to a finite number anyway](https://play.rust-lang.org/?gist=cbd379ad56f76762f177d141ebc1fd80&version=stable&mode=debug).