@dhardy 2018-05-30
Amended.

@MaximoB that sounds worth looking into. I don't think this is the only distribution which could be sped up significantly using a different approach; see #257. If you're interested then please open a PR.